requirements.txt
# Core dependencies
grpcio==1.54.0
grpcio-tools==1.54.0
pandas==2.0.1
numpy==1.24.3
pyyaml==6.0
pytz==2023.3

# Technical analysis
pandas-ta==0.3.14b0
ta-lib==0.4.26

# Data handling and visualization
matplotlib==3.7.1
seaborn==0.12.2

# Interactive Brokers API
ibapi==9.81.1

# Utilities
python-dateutil==2.8.2
requests==2.29.0
tqdm==4.65.0 